20.02.07) Refactor step 1~2 of monater race with Hounx's feedback
Changing package name to monsterRace. Giving responsibility of tryForMove() to monster.
