+++ src/monsterRace/Input.java
@@ -1,4 +1,4 @@
-package JackStep1to2;
+package monsterRace;
 
 import java.util.Scanner;
