+++ src/monsterRace/Output.java
@@ -1,4 +1,4 @@
-package JackStep1to2;
+package monsterRace;
 
 import java.util.List;
